+++ cmd/gnoblog-cli/post_test.go
@@ -1,0 +1,1 @@
+package gnoblog_cli
