--- gnoblog-cli/cmd/errors.go
@@ -1,8 +1,0 @@
-package main
-
-import "errors"
-
-var (
-	errInvalidNumberOfArgs = errors.New("please input only one argument")
-	errInvalidPath         = errors.New("cannot open specified path")
-)
